local_version/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Babel==4.0.0
python-dotenv==1.0.0
dnspython==2.4.2
pytz==2023.3
schedule
Faker
